Folder Restructure for autoloader
Restructuring library for a new auto-loader system that auto loads all
of the scripts into Mari. Added in new tag's into the xml files to allow
the subcategories with the autoloader script. Rearranged the docs folder
to match the Ideascale counter part and updated the sdk files with
information on the library for development purposes.